fix: change startup order for handlers
